REF: runtime -> runtime_metadata module rename.
