+++ okonomiyaki/file_formats/__init__.py
@@ -10,7 +10,7 @@ from ._egg_info import (
     split_egg_name
 )
 from ._package_info import PackageInfo
-from .runtime import (
+from .runtime_metadata import (
     RuntimeVersion, is_runtime_path_valid, runtime_metadata_factory
 )
 from .egg import EggBuilder, EggRewriter
